+++ scripts/cbalance.sh
@@ -1,0 +1,1 @@
+cardano-cli query utxo $TESTNET --address $(cat ../output/$1-gbte-v2-with-bounty-hash2.addr)