+++ changelog.md
@@ -1,3 +1,18 @@
+## 2022-11-30
+#### In `GBTE.TreasuryValidator`:
+- Add check bounty details and output datum to escrow are the same
+#### In `GBTE.EscrowValidator`:
+- Add `escrowValidatorHash` function  
+- Add check for update datum ada/gimbal/deadline is greater or equal 
+- Change `outputFulFillsValue` is same as in new datum
+#### In `GBTE.Compiler`:
+- Change bountyContractHash from string to function (get rid of cardano-cli transaction policyid)
+#### In `Script`
+- Add minimal bash scripts for testing datum-logic 
+  - no reference scripts 
+  - no copy paste utxos
+  - lockTreasury.sh/commitBounty.sh/updateBounty.sh/unlockAllTreasury.sh/balance.sh/cbalance.sh/getTxFunc.sh
+
 ## 2022-11-04
 - Change project name to GPTE
 #### In `GBTE.TreasuryValidator`: